data/image_model.py
# filename: backend/image_model.py
import torch
from torchvision import models, transforms
from PIL import Image, ImageStat
import requests
from io import BytesIO

print("Loading ResNet-50 Model...")
from torchvision.models import ResNet50_Weights
model = models.resnet50(weights=ResNet50_Weights.DEFAULT)
model.eval()

preprocess = transforms.Compose([
    transforms.Resize(256),
    transforms.CenterCrop(224),
    transforms.ToTensor(),
    transforms.Normalize(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225]),
])

def analyze_image(image_url):
    try:
        if not image_url or image_url.startswith("data:"):
            return 0.0, "fake"

        response = requests.get(image_url, timeout=5)
        img = Image.open(BytesIO(response.content)).convert("RGB")

        # CHECK 1: Watermarks/Metadata
        info = img.info
        if info:
            meta = str(info).lower()
            if any(x in meta for x in ["ai generated", "dall-e", "midjourney"]):
                 print("[Image] Found AI Metadata -> Forcing AI_GENERATED")
                 return 0.90, "ai_generated"

        # CHECK 2: Deep Learning Confidence
        img_t = preprocess(img)
        batch_t = torch.unsqueeze(img_t, 0)
        with torch.no_grad():
            out = model(batch_t)
        probabilities = torch.nn.functional.softmax(out, dim=1)[0]
        top_prob = float(probabilities.max().item())
        
        # Threshold set to 0.40 for demo stability on random images
        if top_prob > 0.40:
            return round(top_prob, 2), "fact"
        else:
            # Only flag as fake if the model is very confused
            return round(top_prob, 2), "fake"

    except Exception as e:
        print(f"Image Error: {e}")
        return 0.50, "fact"