data/requirements.txt
flask
flask-cors
torch
torchvision
transformers
sentence-transformers
Pillow
numpy